requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
discord.py
discordbot.py==0.2.3a3
dnspython==2.1.0
idna==3.2
multidict==5.1.0
Pillow==9.2.0
psutil==5.8.0
pymongo==3.12.0
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl
python-dotenv~=0.21.0